requirements.txt
PyYAML==5.3.1
selenium==3.141.0
urllib3==1.25.9
